requirement.txt
aniso8601==9.0.1
bcrypt==3.2.2
boto3==1.24.35
botocore==1.27.35
cffi==1.15.1
click==8.1.3
cryptocode==0.1
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mongoengine==0.24.2
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.4.0
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.10
uuid==1.30
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.1
